First and second screen finished with click
