+++ app/src/main/AndroidManifest.xml
@@ -20,6 +20,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity android:name="com.example.SecondActivity"/>
 
     </application>
