+++ app/src/main/java/com/example/Endpoint.kt
@@ -1,15 +1,19 @@
 package com.example
 
 import com.example.model.mainModel.MainReposResponse
-import com.example.model.secondModel.SecondResponse
+import com.example.model.secondModel.ResponseItem
 import retrofit2.http.GET
+import retrofit2.http.Path
 
 interface Endpoint {
 
     @GET("search/repositories?q=language:Java&sort=stars&page=1")
     suspend fun getPosts(): MainReposResponse
 
-    @GET("repos/kdn251/interviews/pulls")
-    suspend fun getMerge(): SecondResponse
+    @GET("repos/{login}/{name}/pulls")
+    suspend fun getMerge(
+        @Path("login") login: String,
+        @Path("name") repoName: String
+    ): List<ResponseItem>
 
 }
