+++ app/src/main/java/com/example/MainActivity.kt
@@ -1,18 +1,17 @@
 package com.example
 
-import android.content.Intent
 import android.os.Bundle
-import android.widget.Button
 import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.RecyclerView
 import com.example.desafio_android.R
+import com.example.model.mainModel.Items
 import com.example.model.mainModel.MainReposResponse
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
 class MainActivity : AppCompatActivity() {
 
-    val mainViewModel: MainViewModel by viewModel()
+    private val mainViewModel: MainViewModel by viewModel()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -28,10 +27,14 @@ class MainActivity : AppCompatActivity() {
 
     private fun setupList(reposResponse: MainReposResponse) {
         val recyclerView = findViewById<RecyclerView>(R.id.mainRecycler)
-        recyclerView.adapter = MainAdapter(reposResponse.items, this)
+        recyclerView.adapter = MainAdapter(reposResponse.items, this, ::onItemClick)
         recyclerView.addItemDecoration(
             DividerItemDecoration(this, DividerItemDecoration.VERTICAL)
         )
+    }
+
+    fun onItemClick(item : Items){
+        startActivity(SecondActivity.newInstance(this, item.owner.login, item.name))
     }
 
 }
