+++ app/src/main/java/com/example/MainAdapter.kt
@@ -7,7 +7,11 @@ import androidx.recyclerview.widget.RecyclerView.Adapter
 import com.example.desafio_android.R
 import com.example.model.mainModel.Items
 
-class MainAdapter(private val items: List<Items>, private val context: Context) :
+class MainAdapter(
+    private val items: List<Items>,
+    private val context: Context,
+    private val onClick: (Items) -> Unit
+) :
     Adapter<MainViewHolder>() {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): MainViewHolder {
@@ -17,9 +21,8 @@ class MainAdapter(private val items: List<Items>, private val context: Context) 
 
     override fun onBindViewHolder(holder: MainViewHolder, position: Int) {
         val item = items[position]
-        holder.let {
-            holder.bind(item)
-        }
+        holder.bind(item, onClick)
+
     }
 
     override fun getItemCount(): Int {
