+++ app/src/main/java/com/example/MainViewHolder.kt
@@ -3,6 +3,7 @@ package com.example
 import android.view.View
 import android.widget.ImageView
 import android.widget.TextView
+import androidx.constraintlayout.solver.state.State
 import androidx.recyclerview.widget.RecyclerView
 import com.example.desafio_android.R
 import com.example.model.mainModel.Items
@@ -17,9 +18,11 @@ class MainViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
     val stargazer_count = itemView.findViewById<TextView>(R.id.stargazer_count)
     val username = itemView.findViewById<TextView>(R.id.user)
     val avatar = itemView.findViewById<ImageView>(R.id.imageView2)
+    val container = itemView.findViewById<View>(R.id.main_line_view)
 
-    fun bind(item: Items) {
+    fun bind(item: Items, onCLick: (Items) -> Unit) {
 
+        container.setOnClickListener { onCLick(item) }
         title.text = item.name
         description.text = item.description
         forks_count.text = item.forks_count.toString()
