+++ app/src/main/java/com/example/model/mainModel/MainReposResponse.kt
@@ -16,15 +16,6 @@ For support, please feel free to contact me at https://www.linkedin.com/in/syeda
 
 data class MainReposResponse(
 
-    @SerializedName("pulls_url") val pull: String,
-    @SerializedName("forks") val forks: Int,
-    @SerializedName("title") val title: String,
-    @SerializedName("name") val name: String,
-    @SerializedName("description") val description: String,
-    @SerializedName("login") val login: String,
-    @SerializedName("full_name") val full_name: String,
-    @SerializedName("forks_count") val forks_count: Int,
-    @SerializedName("stargazers_count") val stargazers_count: Int,
     @SerializedName("items") val items: List<Items>
 
 )
