+++ app/src/main/java/com/example/model/secondModel/Base.kt
@@ -1,0 +1,21 @@
+package com.example.model.secondModel
+
+import com.google.gson.annotations.SerializedName
+
+data class Base(
+
+    @field:SerializedName("ref")
+    val ref: String? = null,
+
+    @field:SerializedName("repo")
+    val repo: Repo? = null,
+
+    @field:SerializedName("label")
+    val label: String? = null,
+
+    @field:SerializedName("sha")
+    val sha: String? = null,
+
+    @field:SerializedName("user")
+    val user: User? = null
+)
