+++ app/src/main/java/com/example/model/secondModel/Commits.kt
@@ -1,0 +1,9 @@
+package com.example.model.secondModel
+
+import com.google.gson.annotations.SerializedName
+
+data class Commits(
+
+    @field:SerializedName("href")
+    val href: String? = null
+)
