+++ app/src/main/java/com/example/model/secondModel/License.kt
@@ -1,0 +1,21 @@
+package com.example.model.secondModel
+
+import com.google.gson.annotations.SerializedName
+
+data class License(
+
+    @field:SerializedName("name")
+    val name: String? = null,
+
+    @field:SerializedName("spdx_id")
+    val spdxId: String? = null,
+
+    @field:SerializedName("key")
+    val key: String? = null,
+
+    @field:SerializedName("url")
+    val url: String? = null,
+
+    @field:SerializedName("node_id")
+    val nodeId: String? = null
+)
