+++ app/src/main/java/com/example/model/secondModel/SecondRequest.kt
@@ -1,0 +1,5 @@
+package com.example.model.secondModel
+
+import java.io.Serializable
+
+class SecondRequest (val login:String, val name:String) : Serializable
