+++ app/src/main/java/com/example/model/secondModel/Statuses.kt
@@ -1,0 +1,9 @@
+package com.example.model.secondModel
+
+import com.google.gson.annotations.SerializedName
+
+data class Statuses(
+
+    @field:SerializedName("href")
+    val href: String? = null
+)
