+++ app/src/main/res/layout/main_line_view.xml
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/main_line_view"
     android:layout_width="match_parent"
     android:layout_height="170dp">
@@ -97,17 +96,5 @@
         app:layout_constraintStart_toEndOf="@+id/stargazer_count"
         app:layout_constraintTop_toTopOf="@+id/imageView3"
         app:srcCompat="@drawable/likes" />
-
-    <Button
-        android:id="@+id/button"
-        android:layout_width="409dp"
-        android:layout_height="168dp"
-        android:background="#00FBFBFB"
-        android:backgroundTint="#FFFFFF"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
 
 </androidx.constraintlayout.widget.ConstraintLayout>
